+++ src/patch/core/Global.zig
@@ -122,17 +122,15 @@ pub var GLOBAL_FUNCTION: GlobalFunction = .{
 const style_practice_label = rt.MakeTextHeadStyle(.Default, true, .Yellow, .Right, .{rto.ToggleShadow}) catch "";
 
 fn DrawMenuPracticeModeLabel() void {
-    if (GLOBAL_STATE.practice_mode) {
-        _ = GLOBAL_FUNCTION.GDrawText(
-            .Default,
-            rt.MakeText(640 - 20, 16, "Practice Mode", .{}, 0xFFFFFFFF, style_practice_label) catch null,
-        );
-    }
+    _ = GLOBAL_FUNCTION.GDrawText(
+        .SystemP,
+        rt.MakeText(640 - 20, 16, "Practice Mode", .{}, 0xFFFFFFFF, style_practice_label) catch null,
+    );
 }
 
 fn DrawVersionString() void {
     _ = GLOBAL_FUNCTION.GDrawText(
-        .Default,
+        .System,
         rt.MakeText(36, 480 - 24, "{s}", .{VERSION_STR}, 0xFFFFFFFF, null) catch null,
     );
 }