+++ src/patch/core/Toast.zig
@@ -97,6 +97,6 @@ pub fn EarlyEngineUpdateA(gs: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
         }
         const color: u32 = fl.flash_color(item.color, item.timer, ToastSystem.dur_flash) | a;
 
-        _ = gf.GDrawText(.Default, rt.MakeText(16, 8 + y_off + ToastSystem.row_h * i, "{s}", .{item.text}, color, null) catch @panic("failed to draw toast text"));
+        _ = gf.GDrawText(.System, rt.MakeText(16, 8 + y_off + ToastSystem.row_h * i, "{s}", .{item.text}, color, null) catch @panic("failed to draw toast text"));
     }
 }